+++ src/ati/remote/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-from ati.remote.choose import get_remote_func
-
-
-__all__ = ['get_remote_func']